tests: re-ordering tests
[ci skip]
